# applications/app-service/src/main/resources/swagger.yaml
openapi: 3.0.0
info:
  title: Franchise Management API
  version: 1.0.0
  description: API for managing franchises, their branches, and offered products.
tags:
  - name: Franchises
    description: Operations related to franchise management
  - name: Branches
    description: Operations for managing products within a branch
  - name: Products
    description: Operations related to managing branches within a franchise
  - name: Stocks
    description: Operations for updating or retrieving stock information for products.

paths:
  /franchises:
    post:
      tags:
        - Franchises
      summary: Add a new franchise
      operationId: createFranchise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FranchiseRequest'
      responses:
        '201':
          description: Franchise created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FranchiseResponse'
        '400':
          description: If the request data structure is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /franchises/{franchiseId}:
    patch:
      tags:
        - Franchises
      summary: Partial update of a franchise
      operationId: partialUpdateFranchise
      parameters:
        - name: franchiseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FranchiseRequest'
      responses:
        '200':
          description: Franchise updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FranchiseResponse'
        '400':
          description: If the request data structure is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

  /franchises/{franchiseId}/branches:
    post:
      tags:
        - Branches
      summary: Add a new branch to a franchise
      operationId: createBranch
      parameters:
        - name: franchiseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchRequest'
      responses:
        '201':
          description: The branch has been successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchResponse'
        '400':
          description: If the request data structure is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /franchises/{franchiseId}/branches/{branchId}:
    patch:
      tags:
        - Branches
      summary: Partial update of a branch
      operationId: partialUpdateBranch
      parameters:
        - name: franchiseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 123
        - name: branchId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchRequest'
      responses:
        '200':
          description: Branch successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchResponse'
        '400':
          description: If the request data structure is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /franchises/{franchiseId}/branches/{branchId}/products:
    post:
      tags:
        - Products
      summary: Add a new product to a branch
      operationId: createProduct
      parameters:
        - name: franchiseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 123
        - name: branchId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: The product has been successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: If the request data structure is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

  /franchises/{franchiseId}/branches/{branchId}/products/{productId}:
    patch:
      tags:
        - Products
      summary: Partial update of a product
      operationId: partialUpdateProduct
      parameters:
        - name: franchiseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 123
        - name: branchId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 123
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Product successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: The request data structure is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'

  /franchises/{franchiseId}/branches/{branchId}/products/{productId}/stocks:
    patch:
      tags:
        - Stocks
      summary: Update stock of a product
      operationId: updateStock
      parameters:
        - name: franchiseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 123
        - name: branchId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 123
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stock:
                  type: integer
              required:
                - stock
      responses:
        '200':
          description: Stock updated

  /franchises/{franchiseId}/branches/products/stocks/top:
    get:
      tags:
        - Stocks
      summary: Get top-stocked product per branch for a franchise
      operationId: getProductWithMaxStock
      parameters:
        - name: franchiseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 123
      responses:
        '200':
          description: List of top-stocked products by branch
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TopStockProductResponse'

components:
  schemas:
    FranchiseRequest:
      type: object
      properties:
        name:
          type: string
          description: Name of the franchise
          example: Subway
      required:
        - name

    FranchiseResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 123
        name:
          type: string
          description: Name of the franchise

    BranchRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name

    BranchResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 123
        name:
          type: string

    ProductRequest:
      type: object
      properties:
        name:
          type: string
        stock:
          type: integer
          format: int32
          example: 10

    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 123
        name:
          type: string
      required:
        - id
        - name

    ProductStockResponse:
      type: object
      required:
        - id
        - name
        - stock
      properties:
        id:
          type: integer
          format: int64
          description: Product identifier
          example: 123
        name:
          type: string
          description: Product name
          example: Coca-Cola
        stock:
          type: integer
          format: int32
          description: Available units in stock
          example: 10

    TopStockProductResponse:
      type: object
      properties:
        branchName:
          type: string
          description: Branch name
          example: North Branch
        product:
            $ref: '#/components/schemas/ProductStockResponse'

    ErrorDto:
      type: object
      properties:
        code:
          type: string
          example: "400"
        message:
          type: string
          description: Error message
          example: "Invalid request data"
      required:
        - code
        - message